+++ ase/atoms.py
@@ -20,7 +20,7 @@ from ase.constraints import (FixConstraint, FixBondLengths, FixLinearTriatomic,
                              voigt_6_to_full_3x3_stress,
                              full_3x3_to_voigt_6_stress)
 from ase.data import atomic_masses, atomic_masses_common
-from ase.geometry import wrap_positions, find_mic, get_angles, get_distances
+from ase.geometry import wrap_positions, find_mic, get_angles, get_distances, get_dihedrals
 from ase.symbols import Symbols, symbols2numbers
 from ase.utils import deprecated
 
@@ -1499,40 +1499,48 @@ class Atoms:
         # Move back to the rotation point
         self.positions = np.transpose(rcoords) + center
 
-    def get_dihedral(self, a1, a2, a3, a4, mic=False):
+    def get_dihedral(self, a0, a1, a2, a3, mic=False):
         """Calculate dihedral angle.
 
-        Calculate dihedral angle (in degrees) between the vectors a1->a2
-        and a3->a4.
+        Calculate dihedral angle (in degrees) between the vectors a0->a1
+        and a2->a3.
 
         Use mic=True to use the Minimum Image Convention and calculate the
         angle across periodic boundaries.
         """
+        return self.get_dihedrals([[a0, a1, a2, a3]], mic=mic)[0]
 
-        # vector 1->2, 2->3, 3->4 and their normalized cross products:
-        a = self.positions[a2] - self.positions[a1]
-        b = self.positions[a3] - self.positions[a2]
-        c = self.positions[a4] - self.positions[a3]
+    def get_dihedrals(self, indices, mic=False):
+        """Calculate dihedral angles.
+
+        Calculate dihedral angles (in degrees) between the list of vectors
+        a0->a1 and a2->a3, where a0, a1, a2 and a3 are in each row of indices.
+
+        Use mic=True to use the Minimum Image Convention and calculate the
+        angles across periodic boundaries.
+        """
+        indices = np.array(indices)
+        assert indices.shape[1] == 4
+
+        a0s = self.positions[indices[:, 0]]
+        a1s = self.positions[indices[:, 1]]
+        a2s = self.positions[indices[:, 2]]
+        a3s = self.positions[indices[:, 3]]
+
+        # vectors 0->1, 1->2, 2->3
+        v0 = a1s - a0s
+        v1 = a2s - a1s
+        v2 = a3s - a2s
+
+        cell = None
+        pbc = None
+
         if mic:
-            a, b, c = find_mic([a, b, c], self.cell, self.pbc)[0]
-        bxa = np.cross(b, a)
-        cxb = np.cross(c, b)
-        bxanorm = np.linalg.norm(bxa)
-        cxbnorm = np.linalg.norm(cxb)
-        if bxanorm == 0 or cxbnorm == 0:
-            raise ZeroDivisionError('Undefined dihedral angle')
-        bxa /= bxanorm
-        cxb /= cxbnorm
-        angle = np.vdot(bxa, cxb)
-        # check for numerical trouble due to finite precision:
-        if angle < -1:
-            angle = -1
-        if angle > 1:
-            angle = 1
-        angle = np.arccos(angle) * 180 / pi
-        if np.vdot(bxa, c) > 0:
-            angle = 360 - angle
-        return angle
+            cell = self.cell
+            pbc = self.pbc
+
+        return get_dihedrals(v0, v1, v2, cell=cell, pbc=pbc)
+
 
     def _masked_rotate(self, center, axis, diff, mask):
         # do rotation of subgroup by copying it to temporary atoms object
@@ -1606,26 +1614,25 @@ class Atoms:
     def get_angle(self, a1, a2, a3, mic=False):
         """Get angle formed by three atoms.
 
-        calculate angle in degrees between the vectors a2->a1 and
+        Calculate angle in degrees between the vectors a2->a1 and
         a2->a3.
 
         Use mic=True to use the Minimum Image Convention and calculate the
         angle across periodic boundaries.
         """
-
         return self.get_angles([[a1, a2, a3]], mic=mic)[0]
 
     def get_angles(self, indices, mic=False):
         """Get angle formed by three atoms for multiple groupings.
 
-        calculate angle in degrees between vectors between atoms a2->a1
+        Calculate angle in degrees between vectors between atoms a2->a1
         and a2->a3, where a1, a2, and a3 are in each row of indices.
 
         Use mic=True to use the Minimum Image Convention and calculate
         the angle across periodic boundaries.
         """
-
         indices = np.array(indices)
+        assert indices.shape[1] == 3
 
         a1s = self.positions[indices[:, 0]]
         a2s = self.positions[indices[:, 1]]
@@ -1710,7 +1717,6 @@ class Atoms:
         Use mic=True to use the Minimum Image Convention.
         vector=True gives the distance vector (from a0 to a1).
         """
-
         return self.get_distances(a0, [a1], mic=mic, vector=vector)[0]
 
     def get_distances(self, a, indices, mic=False, vector=False):
@@ -1719,7 +1725,6 @@ class Atoms:
         Use mic=True to use the Minimum Image Convention.
         vector=True gives the distance vector (from a to self[indices]).
         """
-
         R = self.arrays['positions']
         p1 = [R[a]]
         p2 = R[indices]